Refactor mb models in to own directory
