+++ app/exports/models.py
@@ -1,7 +1,7 @@
 from django.core.files import File
 from django.db import models
 from django.core.files.storage import default_storage
-from mb.models import BaseModel
+from mb.models.models import BaseModel
 
 class ExportFile(BaseModel):
     file = models.FileField(default_storage)
