+++ app/exports/query_sets/measurements/base_query.py
@@ -1,7 +1,7 @@
 from django.db.models import Subquery, OuterRef, F, Q, Case, When, CharField, Value
 from django.db.models.functions import Concat
-from mb.models import SourceMeasurementValue
-from mb.models import UnitConversion
+from mb.models.models import SourceMeasurementValue
+from mb.models.models import UnitConversion
 
 
 def base_query(measurement_choices):
