+++ app/scripts/script_test.py
@@ -1,7 +1,7 @@
 # exec(open('scripts\\script_test.py').read())
 # exec(open('scripts/script_test.py').read())
 
-from mb.models import *
+from mb.models.models import *
 import pandas as pd
 import io
